+++ relecov_tools/__main__.py
@@ -35,7 +35,7 @@ stderr = rich.console.Console(
     stderr=True, force_terminal=relecov_tools.utils.rich_force_colors()
 )
 
-__version__ = "1.3.0"
+__version__ = "1.4.0"
 
 
 def run_relecov_tools():